requirements.txt
# Requirements for Cl_Automation project
playwright
pytest
allure-pytest
pytest-playwright
Pillow
reportlab

